bd-face-api/requirements.txt
altgraph==0.15
baidu-aip==2.2.4.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
future==0.16.0
idna==2.6
kiwisolver==1.0.1
macholib==1.9
matplotlib==2.2.2
numpy==1.14.3
opencv-python==3.4.1.15
pefile==2017.11.5
Pillow==5.1.0
PyAudio==0.2.11
pydub==0.22.0
pygame==1.9.3
PyInstaller==3.3.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
